+++ package.json
@@ -13,6 +13,7 @@
     "react": "17.0.2",
     "react-native": "0.67.0",
     "react-native-ble-manager": "^8.0.1",
+    "react-native-linear-gradient": "^2.5.6",
     "react-native-modal": "^13.0.0"
   },
   "devDependencies": {
